Initial commit
A simple simulation of world with animals, plants and a human. This program simulates organisms on a board in a turn-based time. After every organism takes its turn the result of all the actions is shown in console on a NxM board, where N and M are the, respectively, vertical and horizontal sizes of the board. It simulates animals such as: sheep, wolf, turtle, antelope etc and plants such as dandelion or deadly nightshade, some of those organisms have special interactions with other if they collide or if it's theirs turn for action.